Refactor Implementation
