+++ ForumChat/app/src/main/java/vanrrtech/app/forumchat/ForumDetails.kt
@@ -144,36 +144,10 @@ class ForumDetails : Fragment() {
                 val url = URL("https://vanrrbackend.000webhostapp.com/forum_chat_backend/DeleteForum.php")
 
 
-                //Opening the URL using HttpURLConnection
-                val conn = url.openConnection() as HttpURLConnection
+                HTTPRESTClient.getHttpRestClient()?.sendPostRequest(url, stringParam,
+                    mHandler!!, null, ConstantDefine.SHOW_REQUEST_DELETE_FORUM_SUCCEDED)
 
-                conn.requestMethod = "POST"
-                val os: OutputStream = conn.getOutputStream()
-                val writer = BufferedWriter(
-                    OutputStreamWriter(os, "UTF-8")
-                )
-                writer.write(stringParam)
 
-                writer.flush()
-                writer.close()
-                os.close()
-                conn.connect()
-
-                //StringBuilder object to read the string from the service
-                val sb = StringBuilder()
-
-                BufferedReader(
-                    InputStreamReader(conn.inputStream, "utf-8")
-                ).use { br ->
-                    var responseLine: String? = null
-                    while (br.readLine().also { responseLine = it } != null) {
-                        sb.append(responseLine!!.trim { it <= ' ' })
-                    }
-                }
-                val message = mHandler?.obtainMessage(ConstantDefine.SHOW_REQUEST_DELETE_FORUM_SUCCEDED, sb.toString())
-                if (message != null) {
-                    mHandler?.sendMessage(message)
-                }
             } catch (e: Exception) {
                 Log.e("error", "onCreate: " + e.message )
             }
